LitLoot/requirements.txt
flask==3.0.2
openai==1.12.0
sentence-transformers
faiss-cpu
requests
beautifulsoup4
tqdm
python-dotenv==1.0.1
httpx==0.27.2
flask-cors==4.0.0
